feat: check job_hash in docDB; if not exist, add new job

+++ code/run_capsule.py
@@ -7,34 +7,25 @@ import itertools
 import json
 import glob
 from tqdm import tqdm
+import logging
+
+from util.docDB_io import check_if_job_exists, add_jobs_to_docDB
 
 LOCAL_NWB_ROOT = "/root/capsule/data/foraging_nwb_bonsai"
+
+logger = logging.getLogger()
+logging.basicConfig(format="%(asctime)s - %(message)s", 
+                    datefmt="%Y-%m-%d %H:%M:%S",
+                    level=logging.INFO
+)
+logger.addHandler(logging.StreamHandler())
 
 def get_all_nwbs(nwb_root=LOCAL_NWB_ROOT):
     # Use glob to get all nwbs
     return glob.glob(f"{nwb_root}/*.nwb")
 
 def get_all_analysis_specs():
-
     analysis_specs = [
-        {
-            "analysis_name": "MLE fitting",
-            "analysis_ver": "first version @ 0.10.0",
-            "analysis_libs_to_track_ver": ["aind_dynamic_foraging_models"],
-            "analysis_args": {
-                "agent_class": "ForagerQLearning",
-                "agent_kwargs": {
-                    "number_of_learning_rate": 1,
-                    "number_of_forget_rate": 1,
-                    "choice_kernel": "one_step",
-                    "action_selection": "softmax",
-                },
-                "fit_kwargs": {
-                    "DE_kwargs": {"polish": True, "seed": 42},
-                    "k_fold_cross_validation": 2,
-                },
-            },
-        },
         {
             "analysis_name": "MLE fitting",
             "analysis_ver": "first version @ 0.10.0",
@@ -54,18 +45,42 @@ def get_all_analysis_specs():
 def generate_job_json():
     nwbs = get_all_nwbs(LOCAL_NWB_ROOT)
     analysis_specs = get_all_analysis_specs()
-    
-    for nwb, analysis_spec in tqdm(itertools.product(nwbs, analysis_specs), "generate job json"):
-        # Generate job_dict and compute hash
+
+    n_job_specs_all = len(nwbs) * len(analysis_specs)
+    n_already_exists = 0
+    new_job_dicts = []
+
+    for nwb, analysis_spec in tqdm(
+        itertools.product(nwbs, analysis_specs),
+        total=n_job_specs_all,
+        desc="Generating job jsons",
+    ):
+        # -- Generate job_dict and compute hash --
         job_dict = {
             "nwb_name": nwb,
             "analysis_spec": analysis_spec,
         }
         job_hash = hash_dict(json.dumps(job_dict))
-        job_dict["job_hash"] = job_hash  # Add hash to job_dict
-        with open(f"/root/capsule/result/{job_hash}.json", "w") as f:
-            json.dump(job_dict, f, indent=4)
 
+        # -- Check if the job_hash already exists on docDB --
+        if check_if_job_exists(job_hash):
+            n_already_exists += 1
+            continue
+
+        # -- If not, create a new record on docDB and write the job_dict to a json file --
+        job_dict["job_hash"] = job_hash  # Add hash to job_dict
+        new_job_dicts.append(job_dict)
+
+        with open(f"/root/capsule/results/{job_hash}.json", "w") as f:
+            json.dump(job_dict, f, indent=4)
+        
+    # -- Batch add new jobs to docDB --
+    logger.info(f"Adding {len(new_job_dicts)} new jobs to docDB.")
+    add_jobs_to_docDB(new_job_dicts)
+
+    logger.info(
+        f"Processed {n_job_specs_all} job specs, {len(new_job_dicts)} new jobs, {n_already_exists} already exist."
+    )
 
 def hash_dict(job_dict):
     return hashlib.sha256(job_dict.encode("utf-8")).hexdigest()
